Bug Fixed

+++ application/controllers/api/C_rest.php
@@ -2217,13 +2217,14 @@ class C_rest extends CI_Controller {
                                     // save to db_purchasing pr_status
                                         // delete first if exist di pr_status dan pr_status_detail
                                             $G_pr_status = $this->m_master->caribasedprimary('db_purchasing.pr_status','PRCode',$PRCode);
-                                            $ID_pr_status = $G_pr_status[0]['ID'];
-                                            $this->db->where('PRCode',$PRCode);
-                                            $this->db->delete('db_purchasing.pr_status');
+                                            if (count($G_pr_status) > 0) {
+                                                $ID_pr_status = $G_pr_status[0]['ID'];
+                                                $this->db->where('PRCode',$PRCode);
+                                                $this->db->delete('db_purchasing.pr_status');
 
-                                            $this->db->where('ID_pr_status',$ID_pr_status);
-                                            $this->db->delete('db_purchasing.pr_status_detail');
-
+                                                $this->db->where('ID_pr_status',$ID_pr_status);
+                                                $this->db->delete('db_purchasing.pr_status_detail');
+                                            }
 
                                     $dataSave = array(
                                         'PRCode' => $PRCode,
